+++ src/main/java/MathCaptain/weakness/Chat/Chat.java
@@ -1,0 +1,26 @@
+package MathCaptain.weakness.Chat;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Chat {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long messageId;
+
+    private Long groupId;
+    private Long userId;
+
+    @Column(nullable = false, length = 1000)
+    private String message;
+
+    @Column(nullable = false)
+    private LocalDateTime timestamp;
+}
