+++ src/main/java/MathCaptain/weakness/Login/LoginController.java
@@ -1,0 +1,15 @@
+package MathCaptain.weakness.Login;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+@RequiredArgsConstructor
+public class LoginController {
+
+//    @GetMapping("/login")
+    public String login() {
+        return "login/loginForm";
+    }
+}
