+++ src/main/java/MathCaptain/weakness/Login/LoginSuccessJWTProvideHandler.java
@@ -1,0 +1,50 @@
+package MathCaptain.weakness.Login;
+
+import MathCaptain.weakness.Security.jwt.JwtService;
+import MathCaptain.weakness.User.Repository.UserRepository;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+
+import java.io.IOException;
+
+@Slf4j
+@RequiredArgsConstructor
+@Transactional
+public class LoginSuccessJWTProvideHandler extends SimpleUrlAuthenticationSuccessHandler {
+
+    private final JwtService jwtService;
+    private final UserRepository userRepository;
+
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response,
+                                        Authentication authentication) throws IOException, ServletException {
+        String email = extractEmail(authentication);
+        String accessToken = jwtService.createAccessToken(email);
+        String refreshToken = jwtService.createRefreshToken();
+
+        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
+        userRepository.findByEmail(email).ifPresent(
+                user -> user.updateRefreshToken(refreshToken)
+        );
+
+        log.info( "로그인에 성공합니다. email: {}" , email);
+        log.info( "AccessToken 을 발급합니다. AccessToken: {}" ,accessToken);
+        log.info( "RefreshToken 을 발급합니다. RefreshToken: {}" ,refreshToken);
+
+        response.getWriter().write("success");
+        // /main 페이지로 리다이렉트 -> 프론트 구현시 연결
+//        response.sendRedirect("/main");
+    }
+
+    private String extractEmail(Authentication authentication) {
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        return userDetails.getUsername();
+    }
+}
