+++ src/main/java/MathCaptain/weakness/Security/jwt/JwtServiceImpl.java
@@ -1,6 +1,6 @@
 package MathCaptain.weakness.Security.jwt;
 
-import MathCaptain.weakness.domain.User.UserRepository;
+import MathCaptain.weakness.User.Repository.UserRepository;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -52,10 +52,12 @@ public class JwtServiceImpl implements JwtService{
     private final ObjectMapper objectMapper;
 
     //== 메서드 ==//
+
+    // AccessToken 생성 (사용자의 email 기반)
     @Override
     public String createAccessToken(String email) {
         return JWT.create()
-                // 빌더를 통해 JWT의 Subject 설정
+                // 빌더를 통해 JWT의 Subject 설정 : AccessToken
                 .withSubject(ACCESS_TOKEN_SUBJECT)
                 // 만료시간 설정
                 .withExpiresAt(new Date(System.currentTimeMillis() + accessTokenValidityInSeconds * 1000))
@@ -65,45 +67,60 @@ public class JwtServiceImpl implements JwtService{
                 .sign(Algorithm.HMAC512(secret));
     }
 
+    // RefreshToken 생성
     @Override
     public String createRefreshToken() {
         return JWT.create()
+                // 빌더를 통해 JWT의 Subject 설정 : RefreshToken
                 .withSubject(REFRESH_TOKEN_SUBJECT)
+                // 만료시간 설정
                 .withExpiresAt(new Date(System.currentTimeMillis() + refreshTokenValidityInSeconds * 1000))
+                // HMA512 알고리즘을 사용하여, secret키로 암호화
                 .sign(Algorithm.HMAC512(secret));
     }
 
+    // RefreshToken 유저에게 Update
     @Override
     public void updateRefreshToken(String email, String refreshToken) {
+        // 해당 email을 가지는 유저 찾기
         usersRepository.findByEmail(email)
                 .ifPresentOrElse(
+                        // 존재하면 refreshToken 업데이트
                         users -> users.updateRefreshToken(refreshToken),
                         () -> new Exception("회원 조회 실패")
                 );
     }
 
+    // RefreshToken 삭제
     @Override
     public void destroyRefreshToken(String email) {
+        // 해당 email을 가지는 유저 찾기
         usersRepository.findByEmail(email)
                 .ifPresentOrElse(
+                        // 존재하면 refreshToken 업데이트
                         users -> users.destroyRefreshToken(),
                         () -> new Exception("회원 조회 실패")
                 );
     }
 
+    // AccessToken과 RefreshToken을 클라이언트에게 전달
     @Override
     public void sendAccessAndRefreshToken(HttpServletResponse response, String accessToken, String refreshToken) {
+        // 응답 메시지의 상태를 200 OK로 설정
         response.setStatus(HttpServletResponse.SC_OK);
 
+        // AccessToken, RefreshToken을 헤더에 담음
         setAccessTokenHeader(response, accessToken);
         setRefreshTokenHeader(response, refreshToken);
 
+        // tokenMap에 두 개의 토큰 내용을 저장
         Map<String, String> tokenMap = new HashMap<>();
         tokenMap.put(ACCESS_TOKEN_SUBJECT, accessToken);
         tokenMap.put(REFRESH_TOKEN_SUBJECT, refreshToken);
 
     }
 
+    // AccessToken을 클라리언트에게 전달
     @Override
     public void sendAccessToken(HttpServletResponse response, String accessToken) {
         response.setStatus(HttpServletResponse.SC_OK);
@@ -114,24 +131,32 @@ public class JwtServiceImpl implements JwtService{
         tokenMap.put(ACCESS_TOKEN_SUBJECT, accessToken);
     }
 
+    // 클라이언트에게서 전달받은 AccessToken을 HTTP 헤더에서 추출
     @Override
     public Optional<String> extractAccessToken(HttpServletRequest request) {
         return Optional.ofNullable(request.getHeader(accessHeader)).filter(
+                // BEARER 접두사로 시작하는 AccessToken을 확인
                 accessToken -> accessToken.startsWith(BEARER)
+                // 찾으면 BEARER 부분을 제외한 AccessToken을 반환
         ).map(accessToken -> accessToken.replace(BEARER, ""));
     }
 
+    // 클라이언트에게서 전달받은 RefreshToken을 HTTP 헤더에서 추출
     @Override
     public Optional<String> extractRefreshToken(HttpServletRequest request) {
         return Optional.ofNullable(request.getHeader(refreshHeader)).filter(
+                // BEARER 접두사로 시작하는 RefreshToken을 확인
                 refreshToken -> refreshToken.startsWith(BEARER)
+                // 찾으면 BEARER 부분을 제외한 RefreshToken을 반환
         ).map(refreshToken -> refreshToken.replace(BEARER, ""));
     }
 
+    // AccessToken에서 email을 claim을 추출하여 맞는지 확인
     @Override
     public Optional<String> extractEmail(String accessToken) {
         try {
             return Optional.ofNullable(
+                    // JWT 라이브러리를 이용해 accessToken 속 email을 추출 (USERNAME_CLAIM)
                     JWT.require(Algorithm.HMAC512(secret)).build().verify(accessToken).getClaim(USERNAME_CLAIM)
                             .asString());
         } catch (Exception e) {
@@ -150,6 +175,7 @@ public class JwtServiceImpl implements JwtService{
         response.setHeader(refreshHeader, refreshToken);
     }
 
+    // 해당 토큰이 유효한지 검사
     @Override
     public boolean isTokenValid(String token) {
         try {
@@ -162,3 +188,4 @@ public class JwtServiceImpl implements JwtService{
     }
 
 }
+
