+++ src/main/java/MathCaptain/weakness/User/DTO/updateUserDto.java
@@ -1,0 +1,15 @@
+package MathCaptain.weakness.User.DTO;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class updateUserDto {
+
+    private String email;
+    private String name;
+    private String nickname;
+    private String phoneNumber;
+
+}
