+++ src/main/java/MathCaptain/weakness/User/DTO/userDto.java
@@ -1,0 +1,16 @@
+package MathCaptain.weakness.User.DTO;
+
+import lombok.*;
+
+@Data
+@Builder(access = AccessLevel.PUBLIC)
+public class userDto {
+
+    private String email;
+    private String password;
+    private String name;
+    private String nickname;
+    private String phoneNumber;
+
+}
+
