--- src/main/java/MathCaptain/weakness/app/Session/SessionConst.java
@@ -1,5 +1,0 @@
-package MathCaptain.weakness.app.Session;
-
-public class SessionConst {
-    public static final String LOGIN_MEMBER = "loginMember";
-}
