--- src/main/java/MathCaptain/weakness/domain/Challenge/Badge.java
@@ -1,16 +1,0 @@
-package MathCaptain.weakness.domain.Challenge;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Badge {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long badgeId;
-
-    @Column(nullable = false, length = 100)
-    private String badgeName;
-}
