--- src/main/java/MathCaptain/weakness/domain/Challenge/Challenge.java
@@ -1,19 +1,0 @@
-package MathCaptain.weakness.domain.Challenge;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Challenge {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long challengeId;
-
-    private Long groupId;
-
-    @Column(nullable = false, length = 500)
-    private String challengeContent;
-}
