--- src/main/java/MathCaptain/weakness/domain/Recruitment/Comment.java
@@ -1,26 +1,0 @@
-package MathCaptain.weakness.domain.Recruitment;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class Comment {
-    @Id
-    @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Long commentID;
-
-    private Long postId;
-
-    private Long userId;
-
-    @Column(nullable = false, length = 1000)
-    private String content;
-
-    @Column(nullable = false)
-    private LocalDateTime commentTime;
-}
