--- src/main/java/MathCaptain/weakness/domain/User/UserDetailsImpl.java
@@ -1,56 +1,0 @@
-package MathCaptain.weakness.domain.User;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class UserDetailsImpl implements UserDetails {
-
-    private final Users users;
-
-    public UserDetailsImpl(Users users) {
-        this.users = users;
-    }
-
-    public Users getUsers() {
-        return users;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return users.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return users.getEmail();
-    }
-
-    public String getEmail() {return users.getEmail();}
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
