--- src/main/java/MathCaptain/weakness/domain/User/UserDetailsServiceImpl.java
@@ -1,22 +1,0 @@
-package MathCaptain.weakness.domain.User;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Users users = userRepository.findByEmail(email)
-                .orElseThrow(() -> new IllegalArgumentException(email));
-        return new UserDetailsImpl(users) {
-        };
-    }
-}
