--- src/main/java/MathCaptain/weakness/domain/User/UserRepository.java
@@ -1,15 +1,0 @@
-package MathCaptain.weakness.domain.User;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<Users, Long> {
-
-    Optional<Users> findByEmail(String email);
-    Optional<Users> findByRefreshToken(String refreshToke);
-
-    boolean existsByEmail(String email);
-    boolean existsByPhoneNumber(String phoneNumber);
-    boolean existsByNickname(String nickname);
-}
