+++ src/main/resources/application.properties
@@ -19,3 +19,5 @@ spring.jpa.show-sql=true
 # JWT
 spring.profiles.include=jwt
 
+logging.level.org.springframework.security=DEBUG
+
